Modification id categorie en Long

+++ src/main/java/fr/eni/buymystuff/dao/DAOArticle.java
@@ -1,7 +1,6 @@
 package fr.eni.buymystuff.dao;
 
 import fr.eni.buymystuff.bo.Categories;
-import org.springframework.data.relational.core.sql.In;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
@@ -22,25 +21,27 @@ public class DAOArticle implements IDAOArticle {
     public DAOArticle(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
+
     @Override
     public void saveArticle(ArticleFormDTO article){
         Long articleId = saveOrUpdateArticle(article);
-        List<Integer> categoryIds = article.getCategories().stream().map(Categories::getId).toList();
+
+        // Maintenant categoryIds est en Long
+        List<Long> categoryIds = article.getCategories().stream()
+                .map(Categories::getId)
+                .toList();
+
         syncArticleCategories(articleId, categoryIds);
     }
 
-
-
     @Override
     public Articles findArticleById(Long id) {
-
         String sql = "SELECT * FROM ARTICLES_VENDUS WHERE no_article = ?";
 
         return jdbcTemplate.queryForObject(
                 sql,
                 (rs, rowNum) -> {
                     Articles article = new Articles();
-
                     article.setId(rs.getLong("no_article"));
                     article.setNomArticle(rs.getString("nom_article"));
                     article.setDescription(rs.getString("description"));
@@ -58,20 +59,17 @@ public class DAOArticle implements IDAOArticle {
 
                     return article;
                 },
-                id // paramètre SQL
+                id
         );
     }
 
     @Override
     public List<Articles> findAllArticlesByUserId(int id) {
-
         String sql = "SELECT * FROM ARTICLES_VENDUS WHERE no_utilisateur = ?";
-
         return jdbcTemplate.query(
                 sql,
                 (rs, rowNum) -> {
                     Articles article = new Articles();
-
                     article.setId(rs.getLong("no_article"));
                     article.setNomArticle(rs.getString("nom_article"));
                     article.setDescription(rs.getString("description"));
@@ -89,9 +87,10 @@ public class DAOArticle implements IDAOArticle {
 
                     return article;
                 },
-                id // paramètre SQL
+                id
         );
     }
+
     public Long saveOrUpdateArticle(ArticleFormDTO article) {
         if (article.getId() == null) {
             // INSERT
@@ -104,17 +103,18 @@ public class DAOArticle implements IDAOArticle {
                 PreparedStatement ps = connection.prepareStatement(insertSql, Statement.RETURN_GENERATED_KEYS);
                 ps.setString(1, article.getNomArticle());
                 ps.setString(2, article.getDescription());
-                ps.setObject(3, article.getDateDebut()); // LocalDateTime → JDBC 4.2 ok
+                ps.setObject(3, article.getDateDebut());
                 ps.setObject(4, article.getDateFin());
                 ps.setDouble(5, article.getPrixInitial());
-                ps.setLong(6, 1); // no_utilisateur, remplacer par l'utilisateur connecté
-                ps.setLong(7, 1); // no_adresse
+                ps.setLong(6, 1); // à remplacer par l'utilisateur connecté
+                ps.setLong(7, 1); // adresse
                 ps.setInt(8, 0);
                 return ps;
             }, keyHolder);
 
             Long generatedId = keyHolder.getKey().longValue();
             article.setId(generatedId);
+
         } else {
             // UPDATE
             String updateSql = "UPDATE ARTICLES_VENDUS SET nom_article=?, description=?, date_debut_encheres=?, date_fin_encheres=?, prix_initial=?, etat_vente=? WHERE no_article=?";
@@ -131,20 +131,21 @@ public class DAOArticle implements IDAOArticle {
         return article.getId();
     }
 
-
-
-    private boolean categoryExists(int categoryId){
+    // Vérifie si la catégorie existe
+    private boolean categoryExists(Long categoryId){
         String sql = "SELECT COUNT(*) FROM CATEGORIES WHERE no_categorie = ?";
         Integer count = jdbcTemplate.queryForObject(sql, Integer.class, categoryId);
         return count != null && count > 0;
     }
-    private void syncArticleCategories(Long articleId, List<Integer> newCategoryIds){
-        // Récupérer les catégories actuelles liées
+
+    // Synchronisation des catégories
+    private void syncArticleCategories(Long articleId, List<Long> newCategoryIds){
+        // Récupérer les catégories actuelles
         String selectSql = "SELECT no_categorie FROM ARTICLES_CATEGORIES WHERE no_article = ?";
         List<Long> existingCategoryIds = jdbcTemplate.queryForList(selectSql, Long.class, articleId);
 
         // Catégories à ajouter
-        List<Integer> toAdd = newCategoryIds.stream()
+        List<Long> toAdd = newCategoryIds.stream()
                 .filter(id -> !existingCategoryIds.contains(id))
                 .toList();
 
@@ -155,7 +156,7 @@ public class DAOArticle implements IDAOArticle {
 
         // INSERT nouvelles catégories
         String insertSql = "INSERT INTO ARTICLES_CATEGORIES(no_article, no_categorie) VALUES (?, ?)";
-        for(int catId : toAdd){
+        for(Long catId : toAdd){
             if(categoryExists(catId)) {
                 jdbcTemplate.update(insertSql, articleId, catId);
             }
